basalam_bot/requirements.txt
fastapi
uvicorn
python-telegram-bot
httpx
python-dotenv
pydantic
